requirements.txt
python-dotenv
tweepy
praw
pytrends
nltk
pandas
matplotlib
jinja2
pdfkit
click
PyYAML
